+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## [Unreleased]
 
+## [0.26.0] - 2023-06-16
+
 ### Fixed
 
-- AirbyteFactory creating duplicated connections
+-   AirbyteFactory creating duplicated connections
 
 ## [0.25.1] - 2023-05-16
 
@@ -259,7 +261,9 @@
 -   Draft of `dp init`, `dp create`, `dp template new`, `dp template list` and `dp dbt`
 -   Draft of `dp compile` and `dp deploy`
 
-[Unreleased]: https://github.com/getindata/data-pipelines-cli/compare/0.25.1...HEAD
+[Unreleased]: https://github.com/getindata/data-pipelines-cli/compare/0.26.0...HEAD
+
+[0.26.0]: https://github.com/getindata/data-pipelines-cli/compare/0.25.1...0.26.0
 
 [0.25.1]: https://github.com/getindata/data-pipelines-cli/compare/0.25.0...0.25.1